+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.1.2 [In development]
 * [ADD] Add size and color parameters to TextDrawing
-* [ADD] Convert WhiteScreen into BlankScreen with a color parameter 
+* [ADD] Convert WhiteScreen into BlankScreen with a color parameter
+* [FIX] Upgrade flutter_blue_plus version
 
 ## 1.1.1
 * [IMP] Integrate font in text control instruction