+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * [ADD] Add a method to enable or disable the display backlight
 * [ADD] Add a method to read the firmware version
 * [ADD] Add the build, help, and camera icons
+* [ADD] Add the chat, nfc, warning, and wifi icons
 
 ## 1.1.1
 * [IMP] Integrate font in text control instruction